db-apis/requirements.txt
flask[async]
flask-cors==4.0.0
boto3==1.34.0
pydantic==2.5.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0